use package style from kscrash
